rename: 독서모임 가입/탈퇴 컴포넌트 명 변경
